+++ include/inviwo/core/properties/ordinalproperty.h
@@ -34,6 +34,8 @@
 #include <inviwo/core/properties/constraintbehavior.h>
 #include <inviwo/core/util/glmvec.h>
 #include <inviwo/core/util/glmmat.h>
+#include <inviwo/core/util/glmmatext.h>
+#include <inviwo/core/util/glmfmt.h>
 
 #include <string>
 #include <type_traits>
