+++ include/inviwo/core/properties/ordinalrefproperty.h
@@ -34,6 +34,8 @@
 #include <inviwo/core/properties/constraintbehavior.h>
 #include <inviwo/core/util/glmvec.h>
 #include <inviwo/core/util/glmmat.h>
+#include <inviwo/core/util/glmmatext.h>
+#include <inviwo/core/util/glmfmt.h>
 #include <inviwo/core/util/assertion.h>
 
 #include <string>
@@ -538,7 +540,7 @@ void OrdinalRefProperty<T>::setMinValue(const T& newMinValue) {
     // Make sure min < max
     modified |= maxValue_.update(glm::max(maxValue_.value, minValue_.value));
 
-    auto val = clamp(get_());  // make sure we clamp after we update min/max
+    auto val = clamp(get_());  // make sure we clamp after we update min/max    
     if (val != get_()) {
         set_(val);
         modified = true;
